Refactors the ips package to interactionpatterns
