+++ mal-impl/src/main/java/esa/mo/mal/impl/interactionpatterns/PubSubIPConsumerHandler.java
@@ -18,18 +18,19 @@
  * limitations under the License. 
  * ----------------------------------------------------------------------------
  */
-package esa.mo.mal.impl.ips;
+package esa.mo.mal.impl.interactionpatterns;
 
 import org.ccsds.moims.mo.mal.MALException;
-import org.ccsds.moims.mo.mal.MALRequestOperation;
+import org.ccsds.moims.mo.mal.MALPubSubOperation;
 import org.ccsds.moims.mo.mal.structures.InteractionType;
 import org.ccsds.moims.mo.mal.transport.MALErrorBody;
 import org.ccsds.moims.mo.mal.transport.MALMessage;
+import org.ccsds.moims.mo.mal.transport.MALMessageHeader;
 
 /**
- * Handles the state machine for a consumer for an REQUEST operation.
+ * Handles the state machine for a consumer for a PUBSUB operation.
  */
-public final class RequestIPConsumerHandler extends SubmitIPConsumerHandler {
+public final class PubSubIPConsumerHandler extends SubmitIPConsumerHandler {
 
     /**
      * Constructor.
@@ -37,12 +38,9 @@ public final class RequestIPConsumerHandler extends SubmitIPConsumerHandler {
      * @param syncOperation true if this is a isSynchronous call.
      * @param responseHolder The response holder.
      */
-    public RequestIPConsumerHandler(final boolean syncOperation,
+    public PubSubIPConsumerHandler(final boolean syncOperation,
             final OperationResponseHolder responseHolder) {
-        super(InteractionType._REQUEST_INDEX,
-                MALRequestOperation._REQUEST_RESPONSE_STAGE,
-                syncOperation,
-                responseHolder);
+        super(InteractionType._PUBSUB_INDEX, 0, syncOperation, responseHolder);
     }
 
     /**
@@ -50,21 +48,35 @@ public final class RequestIPConsumerHandler extends SubmitIPConsumerHandler {
      *
      * @param responseHolder The response holder.
      */
-    public RequestIPConsumerHandler(final OperationResponseHolder responseHolder) {
-        super(InteractionType._REQUEST_INDEX,
-                MALRequestOperation._REQUEST_RESPONSE_STAGE,
-                false,
-                responseHolder);
+    public PubSubIPConsumerHandler(final OperationResponseHolder responseHolder) {
+        super(InteractionType._PUBSUB_INDEX, 0, false, responseHolder);
+    }
+
+    @Override
+    protected boolean checkStage(final int stage) {
+        switch (stage) {
+            case MALPubSubOperation._REGISTER_ACK_STAGE:
+            case MALPubSubOperation._PUBLISH_REGISTER_ACK_STAGE:
+            case MALPubSubOperation._DEREGISTER_ACK_STAGE:
+            case MALPubSubOperation._PUBLISH_DEREGISTER_ACK_STAGE:
+                return true;
+            default:
+                return true;
+        }
     }
 
     @Override
     protected void informListener(final MALMessage msg) throws MALException {
-        if (msg.getHeader().getIsErrorMessage()) {
-            responseHolder.getListener().requestErrorReceived(
-                    msg.getHeader(), (MALErrorBody) msg.getBody(), msg.getQoSProperties());
+        MALMessageHeader header = msg.getHeader();
+
+        if (header.getIsErrorMessage()) {
+            responseHolder.getListener().registerErrorReceived(
+                    header, (MALErrorBody) msg.getBody(), msg.getQoSProperties());
+        } else if ((header.getInteractionStage().getValue() == MALPubSubOperation._PUBLISH_REGISTER_ACK_STAGE)
+                || (header.getInteractionStage().getValue() == MALPubSubOperation._REGISTER_ACK_STAGE)) {
+            responseHolder.getListener().registerAckReceived(header, msg.getQoSProperties());
         } else {
-            responseHolder.getListener().requestResponseReceived(
-                    msg.getHeader(), msg.getBody(), msg.getQoSProperties());
+            responseHolder.getListener().deregisterAckReceived(header, msg.getQoSProperties());
         }
     }
 }
