+++ mal-impl/src/main/java/esa/mo/mal/impl/interactionpatterns/RequestIPProviderHandler.java
@@ -18,21 +18,22 @@
  * limitations under the License. 
  * ----------------------------------------------------------------------------
  */
-package esa.mo.mal.impl.ips;
+package esa.mo.mal.impl.interactionpatterns;
 
 import esa.mo.mal.impl.Address;
 import esa.mo.mal.impl.MALSender;
 import org.ccsds.moims.mo.mal.MALException;
 import org.ccsds.moims.mo.mal.MALInteractionException;
+import org.ccsds.moims.mo.mal.MALRequestOperation;
 import org.ccsds.moims.mo.mal.MOErrorException;
-import org.ccsds.moims.mo.mal.MALSubmitOperation;
-import org.ccsds.moims.mo.mal.provider.MALSubmit;
+import org.ccsds.moims.mo.mal.provider.MALRequest;
+import org.ccsds.moims.mo.mal.transport.MALEncodedBody;
 import org.ccsds.moims.mo.mal.transport.MALMessage;
 
 /**
- * Submit interaction class.
+ * Request interaction class.
  */
-public class SubmitIPProviderHandler extends IPProviderHandler implements MALSubmit {
+public class RequestIPProviderHandler extends IPProviderHandler implements MALRequest {
 
     /**
      * Constructor.
@@ -43,18 +44,24 @@ public class SubmitIPProviderHandler extends IPProviderHandler implements MALSub
      * @throws MALInteractionException if the received message operation is
      * unknown.
      */
-    public SubmitIPProviderHandler(final MALSender sender, final Address address,
+    public RequestIPProviderHandler(final MALSender sender,
+            final Address address,
             final MALMessage msg) throws MALInteractionException {
         super(sender, address, msg);
     }
 
     @Override
-    public MALMessage sendAcknowledgement() throws MALException {
-        return returnResponse(MALSubmitOperation.SUBMIT_ACK_STAGE, (Object[]) null);
+    public MALMessage sendResponse(final Object... result) throws MALInteractionException, MALException {
+        return returnResponse(MALRequestOperation.REQUEST_RESPONSE_STAGE, result);
+    }
+
+    @Override
+    public MALMessage sendResponse(final MALEncodedBody body) throws MALInteractionException, MALException {
+        return returnResponse(MALRequestOperation.REQUEST_RESPONSE_STAGE, body);
     }
 
     @Override
     public MALMessage sendError(final MOErrorException error) throws MALException {
-        return returnError(MALSubmitOperation.SUBMIT_ACK_STAGE, error);
+        return returnError(MALRequestOperation.REQUEST_RESPONSE_STAGE, error);
     }
 }
