+++ mal-impl/src/main/java/esa/mo/mal/impl/interactionpatterns/SubmitIPConsumerHandler.java
@@ -18,7 +18,7 @@
  * limitations under the License. 
  * ----------------------------------------------------------------------------
  */
-package esa.mo.mal.impl.ips;
+package esa.mo.mal.impl.interactionpatterns;
 
 import esa.mo.mal.impl.MALContextFactoryImpl;
 import java.util.Map;
